==> Calib_Summary/Step 3 - Combine Summary.ipynb <==
import pandas as pd
import os

from openpyxl import load_workbook

## Update paths
root_path = os.path.join(os.getcwd())
model_summary = os.path.join(root_path, r"Final_CVM_Model_Summary.xlsx")
survey_summary = os.path.join(root_path, r"Final_CVM_Survey_Summary.xlsx")


calibration_template = os.path.join(root_path, "CVM_Calibration_Template.xlsx")
final_summary = os.path.join(root_path, "CVM_Calibration_FINAL.xlsx")

model_data_workbook = load_workbook(filename=model_summary, read_only=True)
model_data_sheet = model_data_workbook['ModelData']

survey_data_workbook = load_workbook(filename=survey_summary, read_only=True)
survey_data_sheet = survey_data_workbook['SurveyData']

dest_workbook = load_workbook(filename=calibration_template)


dest_model_data_sheet = dest_workbook['ModelData']
dest_model_data_sheet.delete_rows(dest_model_data_sheet.min_row, dest_model_data_sheet.max_row)

dest_survey_data_sheet = dest_workbook['SurveyData']
dest_survey_data_sheet.delete_rows(dest_model_data_sheet.min_row, dest_model_data_sheet.max_row)

for row in model_data_sheet.iter_rows(values_only=True):
    dest_model_data_sheet.append(row)

for row in survey_data_sheet.iter_rows(values_only=True):
    dest_survey_data_sheet.append(row)

# Save changes to the destination workbook
dest_workbook.save(filename=final_summary)

print("complete!")
